JS - entity fichier
