+++ src/Entity/Fichier.php
@@ -1,0 +1,67 @@
+<?php
+
+namespace Entity;
+
+use Symfony\Component\Security\Core\User\UserInterface;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
+
+class Fichier
+
+{
+
+    private $photo;
+    private $fichier;
+
+
+    /**
+     * Gets the value of photo.
+     *
+     * @return mixed
+     */
+    public function getPhoto()
+    {
+        return $this->photo;
+    }
+
+    /**
+     * Sets the value of photo.
+     *
+     * @param mixed $photo the photo
+     *
+     * @return self
+     */
+    private function setPhoto($photo)
+    {
+        $this->photo = $photo;
+
+        return $this;
+    }
+
+    /**
+     * Gets the value of fichier.
+     *
+     * @return mixed
+     */
+    public function getFichier()
+    {
+        return $this->fichier;
+    }
+
+    /**
+     * Sets the value of fichier.
+     *
+     * @param mixed $fichier the fichier
+     *
+     * @return self
+     */
+    private function setFichier($fichier)
+    {
+        $this->fichier = $fichier;
+
+        return $this;
+    }
+
+
+}
+
+?>
